+++ tests/testthat/test_get_type.R
@@ -8,7 +8,7 @@ test_that(
     expect_equal(get_type("GERMAPARLMINI"), "plpr")
     expect_equal(get_type(partition("GERMAPARLMINI", date = "2009-10-28")), "plpr")
     expect_equal(get_type(partition_bundle("GERMAPARLMINI", s_attribute = "date")), "plpr")
-    expect_equal(get_type(Corpus$new("GERMAPARLMINI")), "plpr")
+    expect_equal(get_type(corpus("GERMAPARLMINI")), "plpr")
     expect_equal(get_type(subset("GERMAPARLMINI", date = "2009-11-11")), "plpr")
   }
 )
@@ -19,6 +19,6 @@ test_that(
     expect_equal(is.null(get_type("REUTERS")), TRUE)
     expect_equal(is.null(get_type(partition("REUTERS", places = "kuwait"))), TRUE)
     expect_equal(is.null(get_type(partition_bundle("REUTERS", s_attribute = "places"))), TRUE)
-    expect_equal(is.null(get_type(Corpus$new("REUTERS"))), TRUE)
+    expect_equal(is.null(get_type(corpus("REUTERS"))), TRUE)
   }
 )